Some improvements to web release renamer and command line scraping. Added an error check before deleting nfo when writing out a new one.

+++ JAVMovieScraper/src/moviescraper/doctord/Movie.java
@@ -380,8 +380,9 @@ public class Movie {
 		xml = "<?xml version=\"1.0\" encoding=\"UTF-8\" standalone=\"yes\" ?>"
 				+ "\n" + xml;
 		System.out.println("Xml I am writing to file: \n" + xml);
-
-		nfoFile.delete();
+		
+		if(nfoFile != null && xml != null && xml.length() > 0)
+			nfoFile.delete();
 		FileUtils.writeStringToFile(nfoFile, xml,
 				org.apache.commons.lang3.CharEncoding.UTF_8);
